1.application/cmd.php: refactored  - added ability to run '<section>/<action>' (in addition to 'cmd/<action>')  - added default server variables  - added couple of arguments validations 2.RealtimeController: refactored  - cleanupAction(): session files on disk check refactored  - opentabAction(): refactored to not invoke Indi::cmd() anymore  - closetabAction():    - refactored to not invoke Indi::cmd() anymore    - can now be called only via cmd 3.usage renamed ini()->db->[dbname => name, username => user, password => pass]  - Field_Row->maxLength()  - Indi_Db::factory() 4.Realtime_Row:added:  - onBeforeInsert(): spaceSince setup  - channel(): to create channel nested under session if $this is session  - queue(): to create rabbitmq queue if $this is channel 5.application/ws.php: refactored to be used ONLY for:  - listening for 'queue.deleted'-events coming from rabbitmq    - running cmd('realtime/closetab', [$channel]) to trigger deletion `realtime`-entries of `type`="channel"      - skipped for queues that are not related to current Indi Engine instance 6.Indi.php:  - self::$_rex['cid']  - ws():    - ini('[ws => rabbitmq]')->enabled    - basic_publish(): ini()->db->name prepended to destination queue name  - cmd(): is now alias for global cmd() function 7.Indi_Controller->dispatch(): passed $args as 1st arg to $this->preDispatch() call 8.*_Admin  - ini('ws')->realtime renamed to ini('rabbitmq')->enabled    - preDispatch()    - info()    - deleteContextIfNeed()  - call(): small fix 9.removed *_Cmd->channelAction() 10.*_Row->realtime(): ini('[ws => realtime]')->enabled 11.Indi_Uri_Base->parse($uri):  - added '/admin' to $uri, if need  - prepend '/' to $uri, if not there  - append '/' to $uri, if no query string  - added return $this; 12.deleted library/WsClient.php 13.func.php: added cmd($action, $args = [])

+++ application/cmd.php
@@ -1,21 +1,47 @@
 <?php
-// Setup CMD constant, indicating that this execution was started via Indi::cmd()
+
+// Setup CMD constant, indicating that this execution was started via command line
 define('CMD', true);
 
-// Get method
-$method = $argv[1];
+// Get uri
+if (isset($argv[1])) $uri = $argv[1]; else exit('No command given');
+
+// Get this directory with 'right' slashes
+$_dir_ = str_replace('\\', '/', __DIR__);
+
+// Change current working directory to project root
+chdir(realpath($_dir_ . '/../../../..'));
+
+// Get vendor name
+preg_match('~^(?<DOCUMENT_ROOT>.+?)/(?<VDR>vendor/.+?)/~', $_dir_, $m);
 
 // Get filepath of a temporary file, containing environment variables
-$tmp = $argv[2];
+if (isset($argv[2])) {
 
-// Extract environment variables
-extract(json_decode(file_get_contents($tmp), true));
+    // If 2nd arg is not a path of an existsing file - exit
+    if (!is_file($tmp = $argv[2])) exit("File \"$tmp\" does not exist");
 
-// Delete temporary file
-unlink($tmp);
+    // Extract environment variables
+    extract(json_decode(file_get_contents($tmp = $argv[2]), true));
+
+    // Delete temporary file
+    unlink($tmp);
+}
+
+// Provide default environment variables at least expected by Indi Engine
+$_SERVER += [
+    'DOCUMENT_ROOT' => $m['DOCUMENT_ROOT'],
+    'REQUEST_METHOD' => 'GET',
+    'HTTP_HOST' => 'localhost',
+    'VDR' => $m['VDR'],
+    'STD' => ''
+];
 
 // Boot
-include('index.php');
+include 'index.php';
+
+// If $method is not defined in 'someSection/someAction'-format - assume it's action in CmdController
+if (!preg_match('~^([a-zA-Z0-9]+)/([a-zA-Z0-9]+)$~', $uri)) $uri = 'cmd/' . $uri;
 
 // Dispatch method
-uri()->dispatch((COM ? '' : '/admin') . '/cmd/' . $method . '/', $args);
+uri()->dispatch($uri, $args ?? []);
